# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_internal_domain_models.Notification:
    properties:
      created_at:
        type: string
      data:
        items:
          type: integer
        type: array
      delivered_to:
        items:
          type: integer
        type: array
      id:
        type: integer
      is_read:
        type: boolean
      listing_id:
        type: integer
      message:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: integer
    type: object
  backend_internal_domain_models.NotificationSettings:
    properties:
      created_at:
        type: string
      email_enabled:
        description: Добавляем поле для email
        type: boolean
      notification_type:
        type: string
      telegram_enabled:
        type: boolean
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  backend_pkg_utils.ErrorResponseSwag:
    properties:
      error:
        example: Описание ошибки
        type: string
    type: object
  backend_pkg_utils.SuccessResponseSwag:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  internal_proj_notifications_handler.NotificationMarkReadResponse:
    description: Ответ при изменении статуса уведомления
    properties:
      message:
        description: Сообщение об успешной отметке
        example: notifications.markedAsRead
        type: string
    type: object
  internal_proj_notifications_handler.NotificationSettingsResponse:
    description: Настройки уведомлений пользователя
    properties:
      data:
        description: Массив настроек для разных типов уведомлений
        items:
          $ref: '#/definitions/backend_internal_domain_models.NotificationSettings'
        type: array
    type: object
  internal_proj_notifications_handler.NotificationSettingsUpdateResponse:
    description: Ответ при обновлении настроек уведомлений
    properties:
      message:
        description: Сообщение об успешном обновлении
        example: notifications.settingsUpdated
        type: string
      settings:
        description: Обновленные настройки
        items:
          $ref: '#/definitions/backend_internal_domain_models.NotificationSettings'
        type: array
    type: object
  internal_proj_notifications_handler.PublicEmailSendResponse:
    description: Ответ при отправке email с публичной формы
    properties:
      message:
        description: Сообщение о результате
        example: email.sentSuccessfully
        type: string
      success:
        description: Флаг успешной отправки
        example: true
        type: boolean
    type: object
  internal_proj_notifications_handler.TelegramConnectResponse:
    description: Ответ при успешном подключении Telegram
    properties:
      message:
        description: Сообщение об успешном подключении
        example: telegram.connected
        type: string
    type: object
  internal_proj_notifications_handler.TelegramStatusResponse:
    description: Информация о подключении Telegram к аккаунту
    properties:
      connected:
        description: Подключен ли Telegram
        example: true
        type: boolean
      username:
        description: Имя пользователя в Telegram (если подключен)
        example: john_doe
        type: string
    type: object
  internal_proj_notifications_handler.TelegramTokenResponse:
    description: Токен для связывания аккаунта с Telegram ботом
    properties:
      generated_at:
        description: Время генерации токена
        example: "2024-01-15T10:30:00Z"
        type: string
      token:
        description: Сгенерированный токен
        example: 123_abc...
        type: string
    type: object
  internal_proj_notifications_handler.TelegramWebhookResponse:
    description: Ответ на webhook запрос от Telegram
    properties:
      status:
        description: Статус обработки
        example: OK
        type: string
    type: object
...
  tgbotapi.CallbackQuery:
    properties:
      chat_instance:
        type: string
      data:
        description: optional
        type: string
      from:
        $ref: '#/definitions/tgbotapi.User'
      game_short_name:
        description: optional
        type: string
      id:
        type: string
      inline_message_id:
        description: optional
        type: string
      message:
        allOf:
        - $ref: '#/definitions/tgbotapi.Message'
        description: optional
    type: object
...
  tgbotapi.ChosenInlineResult:
    properties:
      from:
        $ref: '#/definitions/tgbotapi.User'
      inline_message_id:
        type: string
      location:
        $ref: '#/definitions/tgbotapi.Location'
      query:
        type: string
      result_id:
        type: string
    type: object
...
  tgbotapi.InlineQuery:
    properties:
      from:
        $ref: '#/definitions/tgbotapi.User'
      id:
        type: string
      location:
        allOf:
        - $ref: '#/definitions/tgbotapi.Location'
        description: optional
      offset:
        type: string
      query:
        type: string
    type: object
...
  tgbotapi.ShippingQuery:
    properties:
      from:
        $ref: '#/definitions/tgbotapi.User'
      id:
        type: string
      invoice_payload:
        type: string
      shipping_address:
        $ref: '#/definitions/tgbotapi.ShippingAddress'
    type: object
...
  tgbotapi.Update:
    properties:
      callback_query:
        $ref: '#/definitions/tgbotapi.CallbackQuery'
      channel_post:
        $ref: '#/definitions/tgbotapi.Message'
      chosen_inline_result:
        $ref: '#/definitions/tgbotapi.ChosenInlineResult'
      edited_channel_post:
        $ref: '#/definitions/tgbotapi.Message'
      edited_message:
        $ref: '#/definitions/tgbotapi.Message'
      inline_query:
        $ref: '#/definitions/tgbotapi.InlineQuery'
      message:
        $ref: '#/definitions/tgbotapi.Message'
      pre_checkout_query:
        $ref: '#/definitions/tgbotapi.PreCheckoutQuery'
      shipping_query:
        $ref: '#/definitions/tgbotapi.ShippingQuery'
      update_id:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@svetu.rs
    name: API Support
    url: https://svetu.rs/support
  description: API для платформы Sve Tu - маркетплейс объявлений
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://svetu.rs/terms
  title: Sve Tu Marketplace API
  version: "1.0"
paths:
...
  /api/v1/notifications:
    get:
      consumes:
      - application/json
      description: Возвращает список уведомлений пользователя с пагинацией
      parameters:
      - default: 20
        description: Количество записей
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список уведомлений
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_domain_models.Notification'
                  type: array
              type: object
        "500":
          description: notifications.getError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Получить уведомления
      tags:
      - notifications
  /api/v1/notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Отмечает уведомление как прочитанное
      parameters:
      - description: ID уведомления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Уведомление отмечено как прочитанное
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_notifications_handler.NotificationMarkReadResponse'
              type: object
        "400":
          description: validation.invalidNotificationId
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: notifications.updateError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Отметить как прочитанное
      tags:
      - notifications
  /api/v1/notifications/email/public:
    post:
      consumes:
      - application/json
      description: Отправляет email сообщение с публичной формы контактов
      parameters:
      - description: Данные формы
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Email отправлен
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_notifications_handler.PublicEmailSendResponse'
              type: object
        "400":
          description: validation.invalidDataFormat или validation.allFieldsRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: email.sendError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      summary: Отправить email с формы обратной связи
      tags:
      - notifications
  /api/v1/notifications/settings:
    get:
      consumes:
      - application/json
      description: Возвращает настройки уведомлений пользователя для всех типов
      produces:
      - application/json
      responses:
        "200":
          description: Настройки уведомлений
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_notifications_handler.NotificationSettingsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Получить настройки уведомлений
      tags:
      - notifications
    put:
      consumes:
      - application/json
      description: Обновляет настройки уведомлений для конкретного типа
      parameters:
      - description: Настройки
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Настройки обновлены
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_notifications_handler.NotificationSettingsUpdateResponse'
              type: object
        "400":
          description: validation.invalidDataFormat
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: notifications.updateError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Обновить настройки уведомлений
      tags:
      - notifications
  /api/v1/notifications/telegram/connect:
    post:
      consumes:
      - application/json
      description: Связывает аккаунт пользователя с Telegram для получения уведомлений
      parameters:
      - description: Данные для подключения
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Telegram подключен
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_notifications_handler.TelegramConnectResponse'
              type: object
        "400":
          description: validation.invalidDataFormat
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: telegram.connectionError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Подключить Telegram
      tags:
      - notifications
  /api/v1/notifications/telegram/status:
    get:
      consumes:
      - application/json
      description: Проверяет, подключен ли Telegram к аккаунту пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Статус подключения
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_notifications_handler.TelegramStatusResponse'
              type: object
        "401":
          description: auth.unauthorized
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      summary: Статус подключения Telegram
      tags:
      - notifications
  /api/v1/notifications/telegram/token:
    get:
      consumes:
      - application/json
      description: Генерирует токен для связывания аккаунта с Telegram ботом
      produces:
      - application/json
      responses:
        "200":
          description: Токен для подключения
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_notifications_handler.TelegramTokenResponse'
              type: object
        "500":
          description: notifications.tokenGenerateError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Получить токен для Telegram
      tags:
      - notifications
  /api/v1/notifications/telegram/webhook:
    post:
      consumes:
      - application/json
      description: Обрабатывает входящие обновления от Telegram бота
      parameters:
      - description: Telegram update
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/tgbotapi.Update'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_notifications_handler.TelegramWebhookResponse'
              type: object
      summary: Telegram webhook
      tags:
      - notifications
...
securityDefinitions:
  Bearer:
    description: Bearer token для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"